<Enhancement>[PEP 8]: Fix floats
They should be ints

[https://github.com/hasii2011/PyUt/issues/250]

+++ src/org/pyut/ogl/OglPosition.py
@@ -5,5 +5,5 @@ from dataclasses import dataclass
 @dataclass
 class OglPosition:
 
-    x: float = 0.0
-    y: float = 0.0
+    x: int = 0
+    y: int = 0
